backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.36
aiosqlite==0.19.0
asyncpg>=0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]
pydantic-settings
apscheduler==3.10.4
httpx==0.27.2
scikit-learn>=1.4.0
numpy
joblib
python-dotenv==1.0.0
websockets
faker
aiofiles
bcrypt
